share the representer in a module. this means you need two separate representer classes for Create and Update. this is not necessary given that both classes have the same incoming data format.
my solution is to make the representer even more generic, so it really only knows structure plus :populator which is called per incoming property. that way, the representer _instance_ can be shared for different operations (even subclasses) without having to do metaprogramming. this is a fully compositional approach then.

the populator is called on the form, which is responsible for skipping, updating, find_and_create and all the like. this will most probably be done in `Populator` instances that can also be used in "pure" Roar/Representable representers without the form twin.